requirements.txt
# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0

# Redis for caching and session management
redis>=5.0.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Configuration management
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Email validation
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0